refactor: split docs markdown renderers
