+++ crates/ox_content_docs/src/markdown/markdown_html.rs
@@ -8,486 +8,45 @@
 use rustc_hash::FxHashSet;
 
 use super::{
-    clean_summary_text, doc_kind_plural, doc_page_href, effective_members_format,
-    effective_parameters_format, entry_anchor, file_stem, format_count_label, format_kind_label,
-    generate_source_href, get_entry_badges, member_anchor, normalize_signature,
-    parse_example_block, process_doc_text, EntryStats, ExampleBlock, MarkdownDisplayFormat,
-    MarkdownDocsOptions, MarkdownLinkContext, MarkdownPathStrategy, DOC_KIND_ORDER,
+    clean_summary_text, effective_members_format, effective_parameters_format, entry_anchor,
+    format_kind_label, generate_source_href, member_anchor, normalize_signature, process_doc_text,
+    MarkdownDisplayFormat, MarkdownDocsOptions, MarkdownLinkContext, MarkdownPathStrategy,
 };
 use crate::model::{
-    ApiDocEntry, ApiDocMember, ApiDocModule, ApiDocTag, ApiParamDoc, ApiReturnDoc, ApiThrowsDoc,
-    ApiTypeParamDoc,
+    ApiDocEntry, ApiDocMember, ApiDocTag, ApiParamDoc, ApiReturnDoc, ApiThrowsDoc, ApiTypeParamDoc,
 };
 use crate::string_builder::{join3, StringBuilder};
-use rustc_hash::FxHashMap;
 
 mod blocks;
+mod examples;
 mod inline;
+mod modules;
+mod overview;
+mod parameters;
+mod stats;
+mod tags;
+mod throws;
+mod type_parameters;
 
 use blocks::render_markdown_blocks_html;
+use examples::push_examples_html;
+pub(super) use examples::render_module_examples_html;
 use inline::{
     escape_html, render_code_block_html, render_doc_inline_html,
     render_highlighted_inline_code_html, render_inline_html, render_type_inner_html,
 };
-
-pub(super) fn render_details_controls_html(target_selector: &str) -> String {
-    let mut out = StringBuilder::with_capacity(260 + target_selector.len());
-    out.push_str("<div class=\"ox-api-controls\" data-ox-api-target=\"");
-    out.push_str(target_selector);
-    out.push_str("\" role=\"toolbar\" aria-label=\"Reference display controls\">
-<button type=\"button\" class=\"ox-api-controls__button\" data-ox-api-toggle=\"expand\">Open all</button>
-<button type=\"button\" class=\"ox-api-controls__button\" data-ox-api-toggle=\"collapse\">Close all</button>
-</div>");
-    out.into_string()
-}
-
-pub(super) fn render_stats_html(stats: &EntryStats, module_count: Option<usize>) -> String {
-    let mut rendered_items = StringBuilder::new();
-    if let Some(module_count) = module_count {
-        push_stat_html(&mut rendered_items, "modules", module_count, None);
-    }
-
-    push_stat_html(&mut rendered_items, "symbols", stats.entries, None);
-
-    for (index, kind) in DOC_KIND_ORDER.iter().enumerate() {
-        let count = stats.by_kind[index];
-        if count > 0 {
-            push_stat_html(&mut rendered_items, doc_kind_plural(kind), count, None);
-        }
-    }
-
-    if stats.params > 0 {
-        push_stat_html(&mut rendered_items, "parameters", stats.params, None);
-    }
-    if stats.members > 0 {
-        push_stat_html(&mut rendered_items, "members", stats.members, None);
-    }
-    if stats.returns > 0 {
-        push_stat_html(&mut rendered_items, "returns", stats.returns, None);
-    }
-    if stats.examples > 0 {
-        push_stat_html(&mut rendered_items, "examples", stats.examples, None);
-    }
-    if stats.deprecated > 0 {
-        push_stat_html(&mut rendered_items, "deprecated", stats.deprecated, Some("warning"));
-    }
-
-    let rendered_items = rendered_items.into_string();
-    let mut out = StringBuilder::with_capacity(rendered_items.len() + 80);
-    out.push_str("<div class=\"ox-api-stats\" aria-label=\"API reference summary\">\n");
-    out.push_str(&rendered_items);
-    out.push_str("\n</div>");
-    out.into_string()
-}
-
-pub(super) fn render_module_examples_html(examples: &[String]) -> String {
-    if examples.is_empty() {
-        return String::new();
-    }
-
-    let mut examples_html = StringBuilder::new();
-    for (index, example) in examples.iter().enumerate() {
-        if !examples_html.is_empty() {
-            examples_html.push_char('\n');
-        }
-        // Mixed Markdown examples (prose + fenced code) render as HTML blocks; a
-        // pure-code example renders as a single highlighted code block. This avoids
-        // double-wrapping a Markdown example in another `<pre><code>`.
-        let rendered = match parse_example_block(example) {
-            ExampleBlock::Code { code, language } => render_code_block_html(code, language),
-            ExampleBlock::Markdown(markdown) => render_markdown_blocks_html(markdown),
-        };
-        examples_html.push_str(
-            "<div class=\"ox-api-entry__example\">
-<div class=\"ox-api-entry__example-heading\">Example ",
-        );
-        examples_html.push_usize(index + 1);
-        examples_html.push_str(
-            "</div>
-",
-        );
-        examples_html.push_str(&rendered);
-        examples_html.push_str(
-            "
-</div>",
-        );
-    }
-
-    let heading = if examples.len() == 1 { "Example" } else { "Examples" };
-    let mut section = StringBuilder::new();
-    section.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--examples\">
-<h2>",
-    );
-    section.push_str(heading);
-    section.push_str(
-        "</h2>
-",
-    );
-    section.push_str(&examples_html.into_string());
-    section.push_str(
-        "
-</div>
-
-",
-    );
-    section.into_string()
-}
-
-/// Renders module-level lifecycle tags (`@deprecated` / `@experimental`) as a badge
-/// row for the HTML module index (the markdown renderer uses GitHub alerts here).
-/// Returns "" when no lifecycle tags are present.
-pub(super) fn render_module_lifecycle_badges_html(tags: &[ApiDocTag]) -> String {
-    let mut markers = String::new();
-    if tags.iter().any(|tag| tag.tag == "deprecated") {
-        markers.push_str("<span class=\"ox-api-badge ox-api-badge--warning\">deprecated</span>");
-    }
-    if tags.iter().any(|tag| tag.tag == "experimental") {
-        markers.push_str("<span class=\"ox-api-badge ox-api-badge--warning\">experimental</span>");
-    }
-    if markers.is_empty() {
-        return String::new();
-    }
-    join3("<p class=\"ox-api-module__meta\">", &markers, "</p>\n\n")
-}
-
-fn push_stat_html(out: &mut StringBuilder, label: &str, value: usize, tone: Option<&str>) {
-    if !out.is_empty() {
-        out.push_char('\n');
-    }
-    out.push_str("<span class=\"ox-api-stat");
-    if let Some(tone) = tone {
-        out.push_str(" ox-api-stat--");
-        out.push_str(tone);
-    }
-    out.push_str("\">\n  <strong>");
-    out.push_usize(value);
-    out.push_str("</strong>\n  <span>");
-    out.push_str(label);
-    out.push_str("</span>\n</span>");
-}
-
-fn render_entry_badges_html(entry: &ApiDocEntry, class_name: &str) -> String {
-    let badges = get_entry_badges(entry);
-    if badges.is_empty() {
-        return String::new();
-    }
-
-    let mut rendered = StringBuilder::new();
-    for badge in badges {
-        rendered.push_str("<span class=\"ox-api-badge");
-        if let Some(tone) = badge.tone {
-            rendered.push_str(" ox-api-badge--");
-            rendered.push_str(tone);
-        }
-        rendered.push_str("\">");
-        rendered.push_str(&escape_html(&badge.label));
-        rendered.push_str("</span>");
-    }
-
-    let rendered = rendered.into_string();
-    let mut out = StringBuilder::with_capacity(class_name.len() + rendered.len() + 23);
-    out.push_str("<span class=\"");
-    out.push_str(class_name);
-    out.push_str("\">");
-    out.push_str(&rendered);
-    out.push_str("</span>");
-    out.into_string()
-}
-
-fn render_overview_html_item(
-    entry: &ApiDocEntry,
-    href: &str,
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let signature = normalize_signature(entry.signature.as_deref());
-    let summary = clean_summary_text(&process_doc_text(&entry.description, context), 88);
-    let meta = render_entry_badges_html(entry, "ox-api-module__meta");
-    let href = escape_html(href);
-    let heading = if let Some(signature) = signature {
-        let content = render_highlighted_inline_code_html(
-            &signature,
-            "ox-api-module__signature ox-api-module__signature--highlighted",
-            "typescript",
-        );
-        let mut heading = StringBuilder::with_capacity(href.len() + content.len() + 43);
-        heading.push_str("<a href=\"");
-        heading.push_str(&href);
-        heading.push_str("\" class=\"ox-api-module__link\">");
-        heading.push_str(&content);
-        heading.push_str("</a>");
-        heading.into_string()
-    } else {
-        let name = escape_html(&entry.name);
-        let mut heading = StringBuilder::with_capacity(href.len() + name.len() + 80);
-        heading.push_str("<a href=\"");
-        heading.push_str(&href);
-        heading.push_str("\" class=\"ox-api-module__link\"><code class=\"ox-api-module__name\">");
-        heading.push_str(&name);
-        heading.push_str("</code></a>");
-        heading.into_string()
-    };
-
-    let summary_html = if summary.is_empty() {
-        String::new()
-    } else {
-        join3("<span class=\"ox-api-module__summary\">", &render_inline_html(&summary), "</span>")
-    };
-    let kind = escape_html(format_kind_label(&entry.kind));
-    let mut item = StringBuilder::with_capacity(
-        kind.len() + heading.len() + summary_html.len() + meta.len() + 92,
-    );
-    item.push_str("<li><span class=\"ox-api-module__kind\">");
-    item.push_str(&kind);
-    item.push_str("</span><div class=\"ox-api-module__item\">");
-    item.push_str(&heading);
-    item.push_str(&summary_html);
-    item.push_str(&meta);
-    item.push_str("</div></li>");
-    item.into_string()
-}
-
-fn render_params_list_html(
-    params: &[ApiParamDoc],
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let mut rows = StringBuilder::new();
-    for param in params {
-        if !rows.is_empty() {
-            rows.push_char('\n');
-        }
-        let description = render_member_param_description(param);
-        rows.push_str("<li class=\"ox-api-entry__param\">\n  <div class=\"ox-api-entry__param-heading\">\n    <code class=\"ox-api-entry__param-name\">");
-        rows.push_str(&escape_html(&param.name));
-        rows.push_str("</code>\n    <code class=\"ox-api-entry__param-type\">");
-        rows.push_str(&render_type_inner_html(
-            &param.type_annotation,
-            context,
-            &FxHashSet::default(),
-        ));
-        rows.push_str("</code>\n  </div>\n  ");
-        if !description.is_empty() {
-            rows.push_str("<p class=\"ox-api-entry__param-description\">");
-            rows.push_str(&render_doc_inline_html(&description, context));
-            rows.push_str("</p>");
-        }
-        rows.push_str("\n</li>");
-    }
-    let rows = rows.into_string();
-
-    let mut out = StringBuilder::with_capacity(rows.len() + 125);
-    out.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--params\">
-<h4>Parameters</h4>
-<ul class=\"ox-api-entry__params\">
-",
-    );
-    out.push_str(&rows);
-    out.push_str(
-        "
-</ul>
-</div>",
-    );
-    out.into_string()
-}
-
-fn render_params_table_html(
-    params: &[ApiParamDoc],
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let mut rows = StringBuilder::new();
-    for param in params {
-        if !rows.is_empty() {
-            rows.push_char('\n');
-        }
-        let description = render_member_param_description(param);
-        rows.push_str("<tr>\n  <td><code>");
-        rows.push_str(&escape_html(&param.name));
-        rows.push_str("</code></td>\n  <td><code>");
-        rows.push_str(&render_type_inner_html(
-            &param.type_annotation,
-            context,
-            &FxHashSet::default(),
-        ));
-        rows.push_str("</code></td>\n  <td>");
-        rows.push_str(&render_doc_inline_html(&description, context));
-        rows.push_str("</td>\n</tr>");
-    }
-    let rows = rows.into_string();
-
-    let mut out = StringBuilder::with_capacity(rows.len() + 220);
-    out.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--params\">
-<h4>Parameters</h4>
-<table class=\"ox-api-entry__params-table\">
-<thead><tr><th>Name</th><th>Type</th><th>Description</th></tr></thead>
-<tbody>
-",
-    );
-    out.push_str(&rows);
-    out.push_str(
-        "
-</tbody>
-</table>
-</div>",
-    );
-    out.into_string()
-}
-
-fn render_type_parameter_name_html(
-    type_param: &ApiTypeParamDoc,
-    context: Option<&MarkdownLinkContext<'_>>,
-    skip: &FxHashSet<&str>,
-) -> String {
-    let mut name = StringBuilder::new();
-    name.push_str("<code>");
-    name.push_str(&escape_html(&type_param.name));
-    name.push_str("</code>");
-    if let Some(constraint) = &type_param.constraint {
-        name.push_str(" <em>extends</em> <code>");
-        name.push_str(&render_type_inner_html(constraint, context, skip));
-        name.push_str("</code>");
-    }
-    if let Some(default) = &type_param.default {
-        name.push_str(" = <code>");
-        name.push_str(&render_type_inner_html(default, context, skip));
-        name.push_str("</code>");
-    }
-    name.into_string()
-}
-
-/// Names of all type parameters in a list (never linked inside their own siblings'
-/// constraints/defaults).
-fn type_parameter_skip_set(type_parameters: &[ApiTypeParamDoc]) -> FxHashSet<&str> {
-    type_parameters.iter().map(|param| param.name.as_str()).collect()
-}
-
-fn type_parameters_have_descriptions(type_parameters: &[ApiTypeParamDoc]) -> bool {
-    type_parameters.iter().any(|type_param| !type_param.description.trim().is_empty())
-}
-
-fn render_type_parameters_table_html(
-    type_parameters: &[ApiTypeParamDoc],
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let skip = type_parameter_skip_set(type_parameters);
-    let has_description = type_parameters_have_descriptions(type_parameters);
-    let mut rows = StringBuilder::new();
-    for type_param in type_parameters {
-        if !rows.is_empty() {
-            rows.push_char('\n');
-        }
-        rows.push_str("<tr>\n  <td>");
-        rows.push_str(&render_type_parameter_name_html(type_param, context, &skip));
-        if has_description {
-            rows.push_str("</td>\n  <td>");
-            if type_param.description.trim().is_empty() {
-                rows.push_char('-');
-            } else {
-                let description = render_doc_inline_html(&type_param.description, context);
-                if description.is_empty() {
-                    rows.push_char('-');
-                } else {
-                    rows.push_str(&description);
-                }
-            }
-        }
-        rows.push_str("</td>\n</tr>");
-    }
-    let rows = rows.into_string();
-
-    let mut out = StringBuilder::with_capacity(rows.len() + 235);
-    out.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--type-parameters\">
-<h4>Type Parameters</h4>
-<table class=\"ox-api-entry__type-parameters-table\">
-<thead><tr><th>Name</th>",
-    );
-    if has_description {
-        out.push_str("<th>Description</th>");
-    }
-    out.push_str(
-        "</tr></thead>
-<tbody>
-",
-    );
-    out.push_str(&rows);
-    out.push_str(
-        "
-</tbody>
-</table>
-</div>",
-    );
-    out.into_string()
-}
-
-fn render_type_parameters_list_html(
-    type_parameters: &[ApiTypeParamDoc],
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let skip = type_parameter_skip_set(type_parameters);
-    let mut items = StringBuilder::new();
-    for type_param in type_parameters {
-        if !items.is_empty() {
-            items.push_char('\n');
-        }
-        items.push_str("<li class=\"ox-api-entry__type-parameter\">\n  <div class=\"ox-api-entry__type-parameter-heading\">");
-        items.push_str(&render_type_parameter_name_html(type_param, context, &skip));
-        items.push_str("</div>\n  ");
-        if !type_param.description.is_empty() {
-            items.push_str("<p class=\"ox-api-entry__type-parameter-description\">");
-            items.push_str(&render_doc_inline_html(&type_param.description, context));
-            items.push_str("</p>");
-        }
-        items.push_str("\n</li>");
-    }
-    let items = items.into_string();
-
-    let mut out = StringBuilder::with_capacity(items.len() + 145);
-    out.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--type-parameters\">
-<h4>Type Parameters</h4>
-<ul class=\"ox-api-entry__type-parameters\">
-",
-    );
-    out.push_str(&items);
-    out.push_str(
-        "
-</ul>
-</div>",
-    );
-    out.into_string()
-}
-
-fn render_tag_list_html(tags: &[ApiDocTag], context: Option<&MarkdownLinkContext<'_>>) -> String {
-    let mut items = StringBuilder::new();
-    for tag in tags {
-        // Structured tags (lifecycle / since / version) are surfaced as badges, so
-        // exclude them here to avoid duplicating them in the generic tag list.
-        if super::is_structured_tag(&tag.tag) {
-            continue;
-        }
-        items.push_str("<li><span class=\"ox-api-entry__tag-name\">@");
-        items.push_str(&escape_html(&tag.tag));
-        items.push_str("</span><span class=\"ox-api-entry__tag-value\">");
-        items.push_str(&render_doc_inline_html(&tag.value, context));
-        items.push_str("</span></li>");
-    }
-    let items = items.into_string();
-
-    let mut out = StringBuilder::with_capacity(items.len() + 115);
-    out.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--tags\">
-<h4>Tags</h4>
-<ul class=\"ox-api-entry__tags\">",
-    );
-    out.push_str(&items);
-    out.push_str(
-        "</ul>
-</div>",
-    );
-    out.into_string()
-}
+pub(super) use modules::{
+    render_module_index_html, render_module_lifecycle_badges_html, render_module_section_html,
+};
+use overview::render_entry_badges_html;
+use parameters::{render_member_params_html, render_params_list_html, render_params_table_html};
+pub(super) use stats::{render_details_controls_html, render_stats_html};
+use tags::render_tag_list_html;
+use throws::{render_throws_inline_html, render_throws_list_html};
+use type_parameters::{
+    render_member_type_parameters_html, render_type_parameters_list_html,
+    render_type_parameters_table_html,
+};
 
 fn render_member_flags(member: &ApiDocMember) -> String {
     let mut flags = Vec::new();
@@ -625,128 +184,6 @@ fn render_member_default_html(default_value: &str) -> String {
     out.push_str(&escape_html(default_value));
     out.push_str("</code></div>");
     out.into_string()
-}
-
-fn render_member_param_description(param: &ApiParamDoc) -> String {
-    let mut description = param.description.clone();
-    let mut flags = String::new();
-    if param.optional {
-        flags.push_str("optional");
-    }
-    if let Some(default_value) = &param.default_value {
-        if !flags.is_empty() {
-            flags.push_str(" · ");
-        }
-        flags.push_str("default: ");
-        flags.push_str(default_value);
-    }
-    if !flags.is_empty() {
-        if !description.is_empty() {
-            description.push_str(" — ");
-        }
-        description.push_str(&flags);
-    }
-    description
-}
-
-fn render_member_params_html(
-    params: &[ApiParamDoc],
-    options: &MarkdownDocsOptions,
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    if effective_parameters_format(options) == MarkdownDisplayFormat::Table {
-        let mut rows = StringBuilder::new();
-        for param in params {
-            let description = render_member_param_description(param);
-            rows.push_str("<tr><td><code>");
-            rows.push_str(&escape_html(&param.name));
-            rows.push_str("</code></td><td><code>");
-            rows.push_str(&render_type_inner_html(
-                &param.type_annotation,
-                context,
-                &FxHashSet::default(),
-            ));
-            rows.push_str("</code></td><td>");
-            rows.push_str(&render_doc_inline_html(&description, context));
-            rows.push_str("</td></tr>");
-        }
-        let rows = rows.into_string();
-
-        return join3(
-            "<table class=\"ox-api-entry__member-params-table\"><thead><tr><th>Name</th><th>Type</th><th>Description</th></tr></thead><tbody>",
-            &rows,
-            "</tbody></table>",
-        );
-    }
-
-    let mut items = StringBuilder::new();
-    for param in params {
-        let description = render_member_param_description(param);
-        items.push_str("<li><code>");
-        items.push_str(&escape_html(&param.name));
-        items.push_str("</code>");
-        if !description.is_empty() {
-            items.push_char(' ');
-            items.push_str(&render_doc_inline_html(&description, context));
-        }
-        items.push_str("</li>");
-    }
-    join3("<ul class=\"ox-api-entry__member-params\">", &items.into_string(), "</ul>")
-}
-
-fn render_member_type_parameters_html(
-    type_parameters: &[ApiTypeParamDoc],
-    options: &MarkdownDocsOptions,
-    context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let skip = type_parameter_skip_set(type_parameters);
-    if effective_parameters_format(options) == MarkdownDisplayFormat::Table {
-        let has_description = type_parameters_have_descriptions(type_parameters);
-        let mut rows = StringBuilder::new();
-        for type_param in type_parameters {
-            rows.push_str("<tr><td>");
-            rows.push_str(&render_type_parameter_name_html(type_param, context, &skip));
-            if has_description {
-                rows.push_str("</td><td>");
-                if type_param.description.trim().is_empty() {
-                    rows.push_char('-');
-                } else {
-                    let description = render_doc_inline_html(&type_param.description, context);
-                    if description.is_empty() {
-                        rows.push_char('-');
-                    } else {
-                        rows.push_str(&description);
-                    }
-                }
-            }
-            rows.push_str("</td></tr>");
-        }
-        let mut table = StringBuilder::new();
-        table.push_str(
-            "<table class=\"ox-api-entry__type-parameters-table\"><thead><tr><th>Name</th>",
-        );
-        if has_description {
-            table.push_str("<th>Description</th>");
-        }
-        table.push_str("</tr></thead><tbody>");
-        table.push_str(&rows.into_string());
-        table.push_str("</tbody></table>");
-        return table.into_string();
-    }
-
-    let mut items = StringBuilder::new();
-    for type_param in type_parameters {
-        items.push_str("<li class=\"ox-api-entry__type-parameter\"><div class=\"ox-api-entry__type-parameter-heading\">");
-        items.push_str(&render_type_parameter_name_html(type_param, context, &skip));
-        items.push_str("</div>");
-        if !type_param.description.is_empty() {
-            items.push_str("<p class=\"ox-api-entry__type-parameter-description\">");
-            items.push_str(&render_doc_inline_html(&type_param.description, context));
-            items.push_str("</p>");
-        }
-        items.push_str("</li>");
-    }
-    join3("<ul class=\"ox-api-entry__type-parameters\">", &items.into_string(), "</ul>")
 }
 
 fn render_member_table_html(
@@ -1525,74 +962,6 @@ fn push_throws_html(
     );
 }
 
-fn render_throws_list_html(
-    throws: &[ApiThrowsDoc],
-    link_context: Option<&MarkdownLinkContext<'_>>,
-    class_name: &str,
-) -> String {
-    let mut items = StringBuilder::new();
-    for throws_doc in throws {
-        let item = render_throws_item_html(throws_doc, link_context);
-        if item.is_empty() {
-            continue;
-        }
-        items.push_str("<li>");
-        items.push_str(&item);
-        items.push_str("</li>");
-    }
-    if items.is_empty() {
-        return String::new();
-    }
-    let mut out = StringBuilder::new();
-    out.push_str("<ul class=\"");
-    out.push_str(class_name);
-    out.push_str("\">");
-    out.push_str(&items.into_string());
-    out.push_str("</ul>");
-    out.into_string()
-}
-
-fn render_throws_inline_html(
-    throws: &[ApiThrowsDoc],
-    link_context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    throws
-        .iter()
-        .map(|throws_doc| render_throws_item_html(throws_doc, link_context))
-        .filter(|item| !item.is_empty())
-        .collect::<Vec<_>>()
-        .join("; ")
-}
-
-fn render_throws_item_html(
-    throws_doc: &ApiThrowsDoc,
-    link_context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let type_annotation =
-        throws_doc.type_annotation.as_deref().map(str::trim).filter(|value| !value.is_empty());
-    let description = throws_doc.description.trim();
-    if type_annotation.is_none() && description.is_empty() {
-        return String::new();
-    }
-
-    let mut out = StringBuilder::new();
-    if let Some(type_annotation) = type_annotation {
-        out.push_str("<code class=\"ox-api-entry__throws-type\">");
-        out.push_str(&render_type_inner_html(type_annotation, link_context, &FxHashSet::default()));
-        out.push_str("</code>");
-        if !description.is_empty() {
-            out.push_str(" <span class=\"ox-api-entry__throws-description\">");
-            out.push_str(&render_doc_inline_html(description, link_context));
-            out.push_str("</span>");
-        }
-    } else {
-        out.push_str("<span class=\"ox-api-entry__throws-description\">");
-        out.push_str(&render_doc_inline_html(description, link_context));
-        out.push_str("</span>");
-    }
-    out.into_string()
-}
-
 fn render_return_members_html(
     members: &[ApiDocMember],
     options: &MarkdownDocsOptions,
@@ -1830,51 +1199,6 @@ fn push_return_member_signature_html(
     out.push_char(';');
 }
 
-/// Appends the examples section, or nothing when empty.
-fn push_examples_html(body: &mut String, examples: &[String]) {
-    if examples.is_empty() {
-        return;
-    }
-    let mut examples_html = StringBuilder::new();
-    for (index, example) in examples.iter().enumerate() {
-        if !examples_html.is_empty() {
-            examples_html.push_char('\n');
-        }
-        // Mixed Markdown examples (prose + fenced code) render as HTML blocks; a
-        // pure-code example renders as a single highlighted code block. This avoids
-        // double-wrapping a Markdown example in another `<pre><code>`.
-        let rendered = match parse_example_block(example) {
-            ExampleBlock::Code { code, language } => render_code_block_html(code, language),
-            ExampleBlock::Markdown(markdown) => render_markdown_blocks_html(markdown),
-        };
-        examples_html.push_str(
-            "<div class=\"ox-api-entry__example\">
-<div class=\"ox-api-entry__example-heading\">Example ",
-        );
-        examples_html.push_usize(index + 1);
-        examples_html.push_str(
-            "</div>
-",
-        );
-        examples_html.push_str(&rendered);
-        examples_html.push_str(
-            "
-</div>",
-        );
-    }
-
-    body.push_str(
-        "<div class=\"ox-api-entry__section ox-api-entry__section--examples\">
-<h4>Examples</h4>
-",
-    );
-    body.push_str(&examples_html.into_string());
-    body.push_str(
-        "
-</div>\n",
-    );
-}
-
 /// Appends the generic tags section, excluding structured tags (lifecycle / since)
 /// which are surfaced as badges instead. Emits nothing when no tags remain.
 fn push_tag_list_html(
@@ -2052,152 +1376,4 @@ pub(super) fn render_overload_body_html(
 
     out.push_str("</div>\n");
     out
-}
-
-pub(super) fn render_module_section_html(
-    doc: &ApiDocModule,
-    options: &MarkdownDocsOptions,
-    file_name: &str,
-    display_name: &str,
-    count_label: &str,
-    link_context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let mut markdown = String::new();
-    markdown.push_str(
-        "<details class=\"ox-api-module\">
-  <summary>
-    <span class=\"ox-api-module__title\"><a href=\"",
-    );
-    markdown.push_str(&escape_html(&doc_page_href(options, file_name, None)));
-    markdown.push_str("\">");
-    markdown.push_str(&escape_html(display_name));
-    markdown.push_str(
-        "</a></span>
-    <span class=\"ox-api-module__count\">",
-    );
-    markdown.push_str(count_label);
-    markdown.push_str(
-        "</span>
-  </summary>
-  <div class=\"ox-api-module__body\">
-    <ul class=\"ox-api-module__list\">
-",
-    );
-
-    for entry in &doc.entries {
-        let href = doc_page_href(options, file_name, Some(&entry_anchor(&entry.name)));
-        markdown.push_str("      ");
-        markdown.push_str(&render_overview_html_item(entry, &href, link_context));
-        markdown.push('\n');
-    }
-
-    markdown.push_str(
-        "    </ul>
-  </div>
-</details>
-
-",
-    );
-
-    markdown
-}
-
-pub(super) fn render_module_index_html(
-    docs: &[ApiDocModule],
-    options: &MarkdownDocsOptions,
-    doc_to_file: Option<&FxHashMap<String, String>>,
-    display_format: MarkdownDisplayFormat,
-    link_context: Option<&MarkdownLinkContext<'_>>,
-) -> String {
-    let items = docs
-        .iter()
-        .map(|doc| {
-            let display_name = file_stem(&doc.file);
-            let mut file_name = display_name.clone();
-
-            if let Some(doc_to_file) = doc_to_file {
-                if let Some(mapped) = doc_to_file.get(&doc.file) {
-                    file_name.clone_from(mapped);
-                }
-            } else if file_name == "index" {
-                file_name = "index-module".to_string();
-            }
-
-            let count_label = format_count_label(doc.entries.len(), "symbol", Some("symbols"));
-            let href = doc_page_href(options, &file_name, None);
-            let summary = clean_summary_text(&process_doc_text(&doc.description, link_context), 88);
-            (display_name, href, count_label, summary)
-        })
-        .collect::<Vec<_>>();
-
-    if display_format == MarkdownDisplayFormat::Table {
-        let mut rows = StringBuilder::new();
-        for (display_name, href, count_label, summary) in &items {
-            if !rows.is_empty() {
-                rows.push_char('\n');
-            }
-            rows.push_str("<tr><td><a href=\"");
-            rows.push_str(&escape_html(href));
-            rows.push_str("\">");
-            rows.push_str(&escape_html(display_name));
-            rows.push_str("</a></td><td>");
-            rows.push_str(&escape_html(count_label));
-            rows.push_str("</td><td>");
-            rows.push_str(&render_inline_html(summary));
-            rows.push_str("</td></tr>");
-        }
-        let rows = rows.into_string();
-
-        let mut out = StringBuilder::with_capacity(rows.len() + 150);
-        out.push_str(
-            "<table class=\"ox-api-modules-table\">
-<thead><tr><th>Module</th><th>Symbols</th><th>Description</th></tr></thead>
-<tbody>
-",
-        );
-        out.push_str(&rows);
-        out.push_str(
-            "
-</tbody>
-</table>
-
-",
-        );
-        return out.into_string();
-    }
-
-    let mut rows = StringBuilder::new();
-    for (display_name, href, count_label, summary) in &items {
-        if !rows.is_empty() {
-            rows.push_char('\n');
-        }
-        rows.push_str("<li><a href=\"");
-        rows.push_str(&escape_html(href));
-        rows.push_str("\">");
-        rows.push_str(&escape_html(display_name));
-        rows.push_str("</a><span class=\"ox-api-module__count\">");
-        rows.push_str(&escape_html(count_label));
-        rows.push_str("</span>");
-        if !summary.is_empty() {
-            rows.push_str("<span class=\"ox-api-module__summary\">");
-            rows.push_str(&render_inline_html(summary));
-            rows.push_str("</span>");
-        }
-        rows.push_str("</li>");
-    }
-    let rows = rows.into_string();
-
-    let mut out = StringBuilder::with_capacity(rows.len() + 40);
-    out.push_str(
-        "<ul class=\"ox-api-modules-list\">
-",
-    );
-    out.push_str(&rows);
-    out.push_str(
-        "
-</ul>
-
-",
-    );
-    out.into_string()
 }
